update name to server.js
new update
